Retocando detalles3

+++ Aplicacion/ClienteApp/src/main/java/org/example/vista/Perfil/PerfilesGUI.java
@@ -91,46 +91,70 @@ public class PerfilesGUI {
             public void actionPerformed(ActionEvent e) {
                 PerfilDto perfil = new PerfilDto();
                 if (!Validator.validarMaximoCaracteres(textNombre.getText(), 20)) {
-                    JOptionPane.showMessageDialog(null, "El nombre no debe tener mas de 20 caracteres");
-                }else if (!Validator.validarMinimoCaracteres(textNombre.getText(), 3)) {
+                    JOptionPane.showMessageDialog(null, "El nombre no debe tener más de 20 caracteres");
+                } else if (!Validator.validarMinimoCaracteres(textNombre.getText(), 3)) {
                     JOptionPane.showMessageDialog(null, "El nombre debe tener al menos 3 caracteres");
                 } else if (Validator.contieneSoloLetras(textNombre.getText())) {
                     JOptionPane.showMessageDialog(null, "El nombre solo debe contener letras");
                 } else {
-                    perfil.setNombrePerfil(textNombre.getText());
-                    perfil.setEstado((Estados) comboEstado.getSelectedItem());
-                    perfilesPermisoBean.crearPerfil(perfil);
-                    actualizarTabla();
+                    int confirmacion = JOptionPane.showConfirmDialog(null, "¿Estás seguro de guardar el perfil?", "Confirmar Guardar", JOptionPane.YES_NO_OPTION);
+
+                    if (confirmacion == JOptionPane.YES_OPTION) {
+                        perfil.setNombrePerfil(textNombre.getText());
+                        perfil.setEstado((Estados) comboEstado.getSelectedItem());
+                        perfilesPermisoBean.crearPerfil(perfil);
+                        JOptionPane.showMessageDialog(null, "Perfil guardado con éxito");
+                        actualizarTabla();
+                        limpiarCampos();
+                    }
                 }
             }
         });
+
         editarPerfilSeleccionadoButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                PerfilDto perfil = perfilesPermisoBean.obtenerPerfil((Long) tablaPerfiles.getValueAt(tablaPerfiles.getSelectedRow(), 0));
-                /*if (Validator.validarMaximoCaracteres(textNombre.getText(), 20)) {
-                    JOptionPane.showMessageDialog(null, "El nombre no debe tener mas de 20 caracteres");
-                }else if (Validator.validarMinimoCaracteres(textNombre.getText(), 2)) {
-                    JOptionPane.showMessageDialog(null, "El nombre debe tener al menos 3 caracteres");
-                } else if (Validator.contieneSoloLetras(textNombre.getText())) {
-                    JOptionPane.showMessageDialog(null, "El nombre solo debe contener letras");
-                } else{}*/
-                //perfil.setNombrePerfil(textNombre.getText());
+                int filaSeleccionada = tablaPerfiles.getSelectedRow();
+                if (filaSeleccionada == -1) {
+                    JOptionPane.showMessageDialog(null, "Por favor, selecciona un perfil para editar.");
+                    return;
+                }
+
+                PerfilDto perfil = perfilesPermisoBean.obtenerPerfil((Long) tablaPerfiles.getValueAt(filaSeleccionada, 0));
+
+                int confirmacion = JOptionPane.showConfirmDialog(null, "¿El nombre del perfil no puede cambiar?, ¿Estás seguro de modificar el perfil?", "Confirmar Modificación", JOptionPane.YES_NO_OPTION);
+
+                if (confirmacion == JOptionPane.YES_OPTION) {
                     perfil.setEstado((Estados) comboEstado.getSelectedItem());
                     perfilesPermisoBean.modificarPerfil(perfil);
-                    JOptionPane.showMessageDialog(null, "Perfil modificado");
+                    JOptionPane.showMessageDialog(null, "Perfil modificado con éxito");
                     actualizarTabla();
+                    limpiarCampos();
+                }
             }
         });
+
         borrarPerfilSeleccionadoButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                PerfilDto perfil = perfilesPermisoBean.obtenerPerfil((Long) tablaPerfiles.getValueAt(tablaPerfiles.getSelectedRow(), 0));
-                perfilesPermisoBean.eliminarPerfil(perfil);
-                JOptionPane.showMessageDialog(null, "Perfil eliminado");
-                actualizarTabla();
+                int filaSeleccionada = tablaPerfiles.getSelectedRow();
+                if (filaSeleccionada == -1) {
+                    JOptionPane.showMessageDialog(null, "Por favor, selecciona un perfil para dar de Baja.");
+                    return;
+                }
+
+                int confirmacion = JOptionPane.showConfirmDialog(null, "¿Estás seguro de dar de baja el perfil?", "Confirmar Baja", JOptionPane.YES_NO_OPTION);
+
+                if (confirmacion == JOptionPane.YES_OPTION) {
+                    PerfilDto perfil = perfilesPermisoBean.obtenerPerfil((Long) tablaPerfiles.getValueAt(filaSeleccionada, 0));
+                    perfilesPermisoBean.eliminarPerfil(perfil);
+                    JOptionPane.showMessageDialog(null, "Perfil eliminado con éxito");
+                    actualizarTabla();
+                    limpiarCampos();
+                }
             }
         });
+
         //Cuando se ahce click en una fila de la tabla
         tablaPerfiles.addMouseListener(new MouseAdapter() {
             @Override
@@ -150,7 +174,6 @@ public class PerfilesGUI {
                 model.setRowCount(0);
                 for (PerfilDto perfil : perfilesPermisoBean.listarPerfilesPorEstado((Estados) filtroEstadoCombo.getSelectedItem())) {
                     model.addRow(new Object[]{perfil.getId(), perfil.getNombrePerfil(), "Ver permisos", perfil.getEstado()});
-                    System.out.println(perfil.getNombrePerfil());
                 }
             }
         });
@@ -161,7 +184,6 @@ public class PerfilesGUI {
                 model.setRowCount(0);
                 for (PerfilDto perfil : perfilesPermisoBean.listarPerfilesPorNombre(filtroNombreCampo.getText())) {
                     model.addRow(new Object[]{perfil.getId(), perfil.getNombrePerfil(), "Ver permisos", perfil.getEstado()});
-                    System.out.println(perfil.getNombrePerfil());
                 }
             }
         });
@@ -176,11 +198,14 @@ public class PerfilesGUI {
         });
     }
 
+    private void limpiarCampos() {
+        textNombre.setText("");
+    }
+
     private void actualizarTabla() {
         model.setRowCount(0);
         for (PerfilDto perfil : perfilesPermisoBean.obtenerPerfiles()) {
             model.addRow(new Object[]{perfil.getId(), perfil.getNombrePerfil(), "Ver permisos", perfil.getEstado()});
-            System.out.println(perfil.getNombrePerfil());
         }
 
     }
